[Kinetics] Move ChebyshevRate to its own header file
