+++ include/cantera/kinetics/Reaction.h
@@ -12,6 +12,7 @@
 #include "cantera/kinetics/ReactionRate.h"
 #include "cantera/kinetics/RxnRates.h"
 #include "cantera/base/Units.h"
+#include "ChebyshevRate.h"
 #include "InterfaceRate.h"
 #include "Custom.h"
